+++ grocery/src/main/java/com/example/grocery/core/security/repository/RefreshTokenRepository.java
@@ -13,5 +13,5 @@ public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long
     Optional<RefreshToken> findByToken(String token);
 
     @Modifying
-    int deleteByUser(User user);
+    Long deleteByUser(User user);
 }
