+++ grocery/src/main/java/com/example/grocery/webApi/requests/individualCustomer/DeleteIndividualCustomerRequest.java
@@ -2,9 +2,6 @@ package com.example.grocery.webApi.requests.individualCustomer;
 
 import jakarta.validation.constraints.Positive;
 
-// import javax.validation.constraints.NotBlank;
-// import javax.validation.constraints.Size;
-
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -17,5 +14,5 @@ import lombok.Setter;
 public class DeleteIndividualCustomerRequest {
 
     @Positive
-    private int id;
+    private Long id;
 }
